# docs/swagger.yaml
definitions:
  models.AddCategoryRequest:
    properties:
      name:
        maxLength: 256
        type: string
    required:
    - name
    type: object
  models.AddCommentRequest:
    properties:
      body:
        type: string
      postId:
        type: integer
    required:
    - body
    - postId
    type: object
  models.CategoryResponse:
    properties:
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
    type: object
  models.Comment:
    properties:
      authorId:
        type: integer
      body:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      postId:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.UpdateCategoryRequest:
    properties:
      name:
        maxLength: 256
        type: string
    required:
    - name
    type: object
  models.UpdateCommentRequest:
    properties:
      body:
        type: string
    required:
    - body
    type: object
...
info:
  contact:
    email: tubagus.sflh@gmail.com
    name: API Support
    url: http://github.com/bagus2x
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /category:
    post:
      description: Creating a new category.
      parameters:
      - description: the body for creating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AddCategoryRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: create a new category.
      tags:
      - Category
  /category/{categoryId}:
    delete:
      description: deleting a new category.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category id
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: delete a category.
      tags:
      - Category
    get:
      description: finding a category by id.
      parameters:
      - description: Category id
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: get a category by id.
      tags:
      - Category
    put:
      description: updating a category.
      parameters:
      - description: the body for updating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Category id
        in: path
        name: categoryId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CategoryResponse'
      summary: update single category.
      tags:
      - Category
  /comment:
    post:
      description: Creating a new comment.
      parameters:
      - description: the body for creating comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.AddCommentRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
      summary: create a new comment.
      tags:
      - Comment
  /comment/{commentId}:
    delete:
      description: deleting a comment by id.
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment id
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: delete single comment.
      tags:
      - Comment
    get:
      description: finding a comment by id.
      parameters:
      - description: Comment id
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      summary: get a comment by id.
      tags:
      - Comment
    put:
      description: updating a comment.
      parameters:
      - description: the body for updating comment
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCommentRequest'
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment id
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
      summary: update single comment.
      tags:
      - Comment
...
  /post/{postId}/comments:
    get:
      description: finding  comment by post id.
      parameters:
      - description: Post id
        in: path
        name: postId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Comment'
            type: array
      summary: get comments by post id.
      tags:
      - Comment
...
swagger: "2.0"